Create Normal.swift
As it turns out, I need a separate class for each ViewController. So
this one is for the "Normal" difficulty quiz.

I've actually done a lot more than just create the file in this commit.
I hooked up the outlets as needed, and I created a struct that
represents each question. Also, I created a function that selects a
question to display.